fixed soft limits (instead of hard/compile-time limits) for automatically parsed request payloads. Also fixed two bugs related to the subject: 1. a www-form-urlencoded request with incorrect content length and the first key was longer than the provided content length, smisk would induce strange errors (because trying to set NULL into a python dict). This was avoided by in those cases setting the value to None. 2. when parsing form data without content length (an extremely unlikely scenario, given that a POST w/o content length is illegal and caught by the host server) the soft size limit would never be applied. This was solved by simple logic and a wee bit of coffee and magic. Also changed from long to long long for some internal functions

+++ src/utils.h
@@ -45,7 +45,7 @@ int smisk_parse_input_data (char *s, const char *separator, int is_cookie_data,
                             PyObject *dict, const char *charset);
 
 /** Read a line from a FCGI stream */
-size_t smisk_stream_readline (char *str, int n, FCGX_Stream *stream);
+int smisk_stream_readline (char *str, int n, FCGX_Stream *stream);
 
 /**
  * Print bytes - unsafe or outside ASCII characters are printed as \xXX
